Fix import from rsvp module
The rsvp module exports a bunch of named exports, but `import RSVP from
"rsvp"` would try to get the default export, which is undefined.

Tested against rsvp master in a sample ES6 app.

This is not exposed in the test suite because the test suite uses
global-module builds of router and rsvp, which are unaffected by this
difference. Once we migrate tests/tests.js to ES6, this will be easier
to test, but for now let's leave it.